requirements.txt
# Core
requests
pandas
psycopg2-binary
sqlalchemy
python-dotenv
tenacity
pydantic

# dbt
dbt-core
dbt-postgres

# Testing & Linting
pytest
pytest-cov
pytest-mock
ruff

# Utils
loguru
typer
rich
